feat(runner): implement test experiment runner
