+++ .gitignore
@@ -1,3 +1,3 @@
 *cache*
 dist
-tests/results/**
+tests/model/results/
