+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Module with tests."""
+"""Module with all tests."""
