+++ tests/model/__init__.py
@@ -1,0 +1,1 @@
+"""Module with performance tests on language models."""
