+++ tests/software/__init__.py
@@ -1,0 +1,1 @@
+"""Module with automated tests of software quality."""
